INCORRECT BEHAVIOR FOR OO:
refactoring hazzard avoidance for OOO. Instead of following the register renaming chain, we rename only once. That removes the while loop that aimed to follow the renaming chain

Additionally, we removed some code redundancy that existed originally. And we modified the readwrite case with a better separation between handling the read dependency and the write dependency

Things to do:
* Register renaming must use a hidden register file. Right now this code does not work
* Fix the design table in the comments above OOO class