Codelet duplication bug error fixes
- The checkIfReady was only checking on
duplicates that had finished. This fixes it to check for end condition
on the original codelet as well

- Codelets need to update their params once they have
renamed registers. Params are the closure of the codelet
that contains all the registers, but these are
different to the operands. Therefore, upon duplication and register name change
the updateCodeletParams method needed to be called

- Avoid duplication of memory codelets. PReviously we were only checking
For EXECUTE_INST, but memory codelets are also
EXECUTE_INST. Now we also check for isMemoryCodelet()

- Removing unnecessary code and cleaning up some messages.