Reimplementing control store unit to use an array as a queue
Previously an std::Vector was being used for the
control store queue
This patch changes this to be an array, and it
also adds all the necessary atomic operations for the modification of the head and the tail.

This tends to be safer than the previous implementation. Although it may not be needed either.