Adding fault injection mechanism
This patch implements 2 fault injection models
poisson and Weibull. It also allows to select the
model through the envvar
SCM_FAULT_INJECTION_MODE that
may take the values:
- POISSON_FAULT_INJECTION
- WEIBULL_FAULT_INJECTION
- NO_FAULT_INJECTION

The models use the execution time of the codelet to determine the
probability of injection of fault

Furthermore, this patch uses the env vars
SCM_LAMBDA and SCM_WEIBULL_BETA to change
the fault injection parameters.

Upon injection, this patch scrambles the registers and also
changes the flag itFailed in the instruction
such that the duplicator can detect faults.

Finally, this patch provides early termination of the program upon unrecoverable error.