Fixing error injection.
The error injection was using the codelet time,
however, this is wrong as it correlates the errors for different CUs.

This patch uses a timer for each CU, and also considers
the time it takes to execute the codelet.
There is a slight error coming from the fact that errors are inejcted only
at beginning or end of a codelet.

This patch also changes the way comparison works, since Comparison was failing
even though we had no error injected. This is likely due to
rounding errors in floating point numbers. Therefore, we just check if the codelet
failed or not