Fixing datarace in the incrementing of tail and head
There was a possible datarace whenever the current++ instruction
was executed and the CU was checking if the queue was empty.

Order would be:
current++
it reaches the end of the q
the CU checks if the head and tail are the same, and they are not
then it tries to execute on nullptr.
Then the producer would reset the current to next