Fixing register renaming
There was an error where register renaming would get the next pointer of the
same register number. This is incorrect as it should allow to search
for more than 1 register. This patch fixes it by splitting getting
the next register number, and the register pointer based on a size and number