Overhaul of the duplication control flow
This removes the READY_DUP, EXECUTING_DUP states
that we previous had.

The problem was twofold:
- Adding elements to the instruction buffer while iterating on it
- In the list of duplicates, the original was being
added during the clean up process.

It is likely the first part had a solution, but it may be cleaner like this
where we do not do add the duplicates to the current buffer window,
but rather keep another one

Jose